feat: Changed Socials links to a dropdown for more consice navbar. Moved FAQ to footer and merged engage links for smaller navbar footprint.
