+++ src/theme/NavbarItem/ComponentTypes.js
@@ -1,7 +1,6 @@
 import ComponentTypes from "@theme-original/NavbarItem/ComponentTypes";
 import AccountButton from '@site/src/components/NavbarItems/AccountButton';
 
-
 export default {
     ...ComponentTypes,
     'custom-accountButton': AccountButton,
